Make the light themes one tone darker
With the added fidelity the new palette has, this allows us to make the light theme a bit more varied.
This also makes the terminal background the same as code has, which makes more sense.

I feel like v0.2 neglected the light theme, so this hopefully is a step in the right direction.